fixed some bugs and starting working on config updates
