+++ src/TodoGenieLib/Functions/FileFunctions.cs
@@ -1,3 +1,6 @@
+using System.Text;
+using System.Text.RegularExpressions;
+using TodoGenieLib.Models;
 using TodoGenieLib.Utils;
 
 namespace TodoGenieLib.Functions;
@@ -55,5 +58,43 @@ public class FileFunctions {
             Files.Add(file);
         }
         return Files;
+    }
+    public static string GetGithubEndpoint(string rootDir) {
+        string endpoint = string.Empty;
+        var content = File.ReadAllLines(Path.Join(rootDir, ".git", "config"));
+        foreach(var line in content) {
+            var match = Regex.Match(line, @"url = https?://github.com/(.+)/(.+).git");
+            if(match.Success) {
+                endpoint = $"/repos/{match.Groups[1].Value}/{match.Groups[2].Value}/issues";
+                break;
+            }
+        }
+        return endpoint;
+    }
+    public static ConfigModel SetupConfigDir(ConfigModel config) {
+        if(!Directory.Exists(config.ConfigDirectory) && !string.IsNullOrEmpty(config.ConfigDirectory)) {
+            Directory.CreateDirectory(config.ConfigDirectory);
+        }
+        string secretFile = Path.Join(config.ConfigDirectory, config.SecretFileName); 
+        if(!File.Exists(secretFile)) {
+            var stream = File.Create(secretFile);
+            string tempKey = string.Empty;
+            while(string.IsNullOrEmpty(tempKey)) {
+                Console.Write("Enter Github Api Key: ");
+                tempKey = Console.ReadLine()!;
+            }
+            
+            config.GithubApiKey = Crypt.Encrypt(tempKey);
+   
+            var contents = new UTF8Encoding(true).GetBytes("{ \"GithubApiKey\": \"" + config.GithubApiKey + "\" }");
+   
+            stream.Write(contents, 0, contents.Length);
+            stream.Close();
+        }    
+        return config;
+    }
+    public static void SetConfig(ConfigModel config) {
+        //TODO: setup saving to config file
+        throw new NotImplementedException();
     }
 }
